Integrated Payment System

+++ ProduktEdycja.cs
@@ -25,6 +25,8 @@ namespace Magazyn_Spedycji
 
         private void Form1_Load(object sender, EventArgs e)
         {
+            id_prod.Hide();
+            label1.Hide();
             dataGridView1.Show();
             OleDbCommand createProdukty = new OleDbCommand();
             createProdukty.Connection = con;
@@ -35,6 +37,7 @@ namespace Magazyn_Spedycji
             produkty.Fill(tabelaProdukty);
             dataGridView1.DataSource = tabelaProdukty;
             con.Close();
+
         }
     
         public void sprawdz_poprawnosc()
@@ -49,7 +52,7 @@ namespace Magazyn_Spedycji
 
             }
       
-            else if (!r_ilosc_id.IsMatch(ilosc_prod.Text) || !r_ilosc_id.IsMatch(id_prod.Text))
+            else if (!r_ilosc_id.IsMatch(ilosc_prod.Text))
             {
                 MessageBox.Show("Niewlasciwe dane w polu!");
             }
@@ -72,7 +75,7 @@ namespace Magazyn_Spedycji
         }
         
 
-        private void edycja_rekordu_Click(object sender, EventArgs e)
+        private void edycja_rekordu_Click(object sender, EventArgs e) 
         {
             
             sprawdz_poprawnosc();
@@ -85,7 +88,7 @@ namespace Magazyn_Spedycji
                 laczenie.Connection = con;
                 // "' ,NazwaProduktu='" + nazwa_prod.Text + "' ,Ops='" + opis_prod + "' ,KosztStandardowy'" + Convert.ToInt32(koszt_prod.Text) + "' ,CenaKatalogowa='" + Convert.ToInt32(cenakat_prod.Text) + "' ,Ilosc='" + Convert.ToInt32(ilosc_prod.Text) +
 
-                string queryEdycja = "update Produkty set KodProduktu='" + kod_prod.Text + "', NazwaProduktu='" + nazwa_prod.Text + "', Ops='" + opis_prod.Text + "', CenaJednostkowa='" + koszt_prod.Text + "', Ilosc='" + ilosc_prod.Text + "' where ID=" + id_prod.Text;
+                string queryEdycja = "update Produkty set KodProduktu='" + kod_prod.Text + "', NazwaProduktu='" + nazwa_prod.Text + "', Opis='" + opis_prod.Text + "', CenaJednostkowa='" + koszt_prod.Text + "', Ilosc='" + ilosc_prod.Text + "' where ID=" + id_prod.Text;
                 laczenie.CommandText = queryEdycja;
                 laczenie.ExecuteNonQuery();
                 MessageBox.Show("Pomyślnie zaktualizowano produkt");
@@ -97,9 +100,6 @@ namespace Magazyn_Spedycji
 
         private void dodanie_rekordu_Click(object sender, EventArgs e)
         {
-
-
-            //insert into Produkty(ID, KodProduktu, NazwaProduktu, Ops, KosztStandardowy, CenaKatologowa, Ilosc) values('3', 'KAC-17', 'Hula hop rozowe', 'super koleczko do krecenia', '25', '30', '7')
        sprawdz_poprawnosc();
 
             if (poprawne == true)
@@ -108,14 +108,11 @@ namespace Magazyn_Spedycji
                 con.Open();
                 OleDbCommand laczenie = new OleDbCommand();
                 laczenie.Connection = con;
-
-                string queryDodaj = "insert into Produkty(ID, KodProduktu, NazwaProduktu, Ops, CenaJednostkowa, Ilosc) values('" + id_prod.Text + "','" + kod_prod.Text + "','" + nazwa_prod.Text + "','" + opis_prod.Text + "','" + koszt_prod.Text + "','" + ilosc_prod.Text + "')";
+                string queryDodaj = "insert into Produkty(KodProduktu, NazwaProduktu, Opis, CenaJednostkowa, Ilosc) values('" + kod_prod.Text + "','" + nazwa_prod.Text + "','" + opis_prod.Text + "','" + koszt_prod.Text + "','" + ilosc_prod.Text + "')";
                 laczenie.CommandText = queryDodaj;
                 laczenie.ExecuteNonQuery();
                 MessageBox.Show("Pomyślnie dodano produkt");
                 con.Close();
-
-     
             }
             odswiez_gridview();
         }
@@ -123,36 +120,38 @@ namespace Magazyn_Spedycji
         private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             kolumna = e.RowIndex;
-            DataGridViewRow kol = dataGridView1.Rows[kolumna];
-            id_prod.Text = kol.Cells[0].Value.ToString();
-            kod_prod.Text = kol.Cells[1].Value.ToString();
-            nazwa_prod.Text = kol.Cells[2].Value.ToString();
-            opis_prod.Text = kol.Cells[3].Value.ToString();
-            koszt_prod.Text = kol.Cells[4].Value.ToString();
-            ilosc_prod.Text = kol.Cells[5].Value.ToString();
-            wycofane_prod.Checked = Convert.ToBoolean( kol.Cells[6].Value.ToString());
+            if (kolumna < 0 || kolumna > 6) MessageBox.Show("Wybierz produkt!");
+            else
+            {
+                DataGridViewRow kol = dataGridView1.Rows[kolumna];
+                id_prod.Text = kol.Cells[0].Value.ToString();
+                kod_prod.Text = kol.Cells[1].Value.ToString();
+                nazwa_prod.Text = kol.Cells[2].Value.ToString();
+                opis_prod.Text = kol.Cells[3].Value.ToString();
+                koszt_prod.Text = kol.Cells[4].Value.ToString();
+                ilosc_prod.Text = kol.Cells[5].Value.ToString();
+            }
+           
         }
-
-
-
         private void usun_rdk_Click(object sender, EventArgs e)
         {
-            kolumna = dataGridView1.CurrentCell.RowIndex;
-            con.Open();
-            OleDbCommand laczenie = new OleDbCommand();
-            laczenie.Connection = con;
-            string queryUsun = "Delete FROM Produkty where ID="+kolumna;
-            laczenie.CommandText = queryUsun;
-            laczenie.ExecuteNonQuery();
-            con.Close();
-
-            odswiez_gridview();
-
-        }
-
-        private void label7_Click(object sender, EventArgs e)
-        {
-
+            DialogResult result = MessageBox.Show("Czy na pewno chcesz usunąć ten produkt?", "Uwaga!", MessageBoxButtons.YesNo);
+            if (result == DialogResult.Yes)
+            {
+                kolumna = dataGridView1.CurrentCell.RowIndex;
+                con.Open();
+                OleDbCommand laczenie = new OleDbCommand();
+                laczenie.Connection = con;
+                string queryUsun = "Delete FROM Produkty where ID=" + kolumna;
+                laczenie.CommandText = queryUsun;
+                laczenie.ExecuteNonQuery();
+                con.Close();
+                odswiez_gridview();
+            } else if (result == DialogResult.No)
+            {
+                
+            }
+        
         }
     }
 }
